refactor(ui): split sound pack manager dialog
Move sound pack manager models, UI construction, mapping and preview handlers, pack actions, and community sharing into focused mixins while preserving the dialog entry point.

Tested: sound pack manager import smoke test; uv run ruff check src tests scripts; uv run pyright

Confidence: high
